Extract user data into separate config file.
The [user] part in a .gitconfig contains a bit more sensitive info; the
e-mail address for instance.
This change extracts that part into a separate config file, which can
then be ignored in the repository.
It has the added bonus that one can change user name and e-mail
depending on the machine we're working on, work/private for example.