Replace home rolled git_branch() with git-prompt.sh
While the git_branch functionality got the job done most of the time,
there was an issue that bothered me; when working with submodules the
prompt would not show anything when entering a directory where a git
submodule resided.  I decided that it wasn't worth the time fixing it,
so I took what the git install already provided (homebrew install).

I also changed the color of the cwd in the prompt for consistency: since
__git_ps1 shows the same colors as `git status -sb`, I've made the
prompt use the same color as `ls` for the directory.