gitconfig: pretty format is now more dark/light background agnostic.
Been switching from light to dark backgrounds in terminal and using
black/white respectively for the commit messages in the pretty format
definition.  This was ridiculous of course, using 'reset' makes this
much more robust when switching and now the pretty format looks okay on
both light and dark terminal backgrounds.