Add 'todo' alias.
The todo file isn't really 'do now' file, more some stuff to keep in
mind or to do at some time, but I do have to keep an eye on it.
This file will also be symlinked as /etc/motd reminding me of stuff
every time I open a terminal.