Set GIT_EDITOR.
Without me noticing, git has been using my std. vim (the supplied with
OSX) because my $EDITOR variable was just 'vim', with no path and git
doesn't have my include path.  But now that Ultisnips (a Vim plugin) has
a dependency on Vim7.4 I noticed Vim spitting out a warning when
committing from the command line.  At first I thought a lot of other
things were to blame (I had installed a few things, and so on), but upon
further investigation, it was the $EDITOR not set correctly.
With this change, $EDITOR environment variable is set and I've set
GIT_EDITOR for good measure to the same value.