Prompt: Maximize line length for typing.
Especially on a laptop screen having editor + terminal open, the prompt
can get quite long, leaving not much place for typing quite simple
commands.  By putting in a new line to the prompt, typing line length is
maximized, and I have some more room in the prompt to add user@host.
The downside is that vertical space is lowered, but that's much less of
a concern.