Add PHP version to the bash prompt.
I'm doing support work on older projects, most of them don't need a full
on Vagrant box to run, but most do need an older PHP version (5 vs. 7
most of the time).
I'm finding myself typing `php --version` a lot to see which PHP version
I'm currently using.  I am not sure yet if this is because I need to
check it or because I just wrote a little script to switch between
versions (see bin/phps in this repository).